MIPS pixelflinger initial release, merged to a single commit
This release has some known bugs, described below. For this reason, the
optimizations are disabled by default. We have plans to fix the bugs,
and continue with further performance optimizations.

The optimizations can be enabled at runtime by setting a property:

    setprop debug.pf.asm 1

they can be disabled at any time by:

    setprop debug.pf.asm 0

To look at the code that is generated, set this property:

    setprop debug.pf.disasm 1

Known bugs
----------

- Some error dialogs gray out the background screen, optimimixzed pixelflinger
incorrectly gives a red cast to this background.

- There are texturing bugs, which show up as blurry/fuzzy rendering on
ApiDemos/Graphics/OpenGL/Sprite Text and shows as a crash on
ApiDemos/Graphics/OpenGL/Textured-Triangle. I suspect most OpenGL
apps that use textures will not work.

- There is a bug in ApiDemos in cupcake, where the
ApiDemos/Graphics/OpenGL/Translucent-GLSurfaceView does not work.
It is fine in donut. This is not related to pixelflinger changes.