[MIPS] Fix rendering of Gallery app background on big-endian.
This fixes a case where 32-bit RGBA pixels were being used as a texture,
and must be byte-swapped on big-endian machine.

Note that the Gallery app still renders the image thumbnails wrong.
At this point, I do not think this is pixelflinger related, but I
cannot prove that yet. The final full-size images are rendered
correctly.

Signed-off-by: Paul Lind <plind@mips.com>