[MIPS] Increase pixelflinger code cache size.
When investigating Gallery3D app issues, I noticed that more code was
generated than would fit in cache, causing good code to be evicted, and
repeated regeneration of the same code stubs.

Increasing the code cache on mips from 12Kbytes to 32 Kbytes.

This visibly increased the performance of the Gallery3D app.

Also, saw that a errant printf() was being caught by the logs as a warning.
Changed this message to use LOGI(), informational.

Change-Id: If6048c7ebe7fe73fdae40ee18816836f76c9ea63

Signed-off-by: Paul Lind <plind@mips.com>